refactor: split service.ts into service/ modules
Extract the monolithic service.ts into focused modules under src/service/:
- service/index.ts (re-export hub)
- service/load-netlist.ts (design loading)
- service/component-grouping.ts (component aggregation)
- service/regex-helpers.ts (input validation)
- service/tools/ (one file per MCP tool handler)
