--- src/service.ts
@@ -1,1139 +1,0 @@
-/**
- * Netlist Service
- *
- * Query methods for Cadence and Altium netlists.
- * All methods take a path to the design file as input (relative or absolute).
- */
-
-import { exec } from "child_process";
-import * as fs from "fs";
-import { tmpdir } from "os";
-import path from "path";
-import { promisify } from "util";
-import { createMutex } from "./async-mutex.js";
-import { discoverDesigns, findHandler, parseDesign } from "./parsers/index.js";
-import { resolvePath } from "./paths.js";
-
-import {
-  naturalSort,
-  traverseCircuitFromNet,
-  computeCircuitHash,
-  matchesRefdesType,
-  getRefdesPrefix,
-  isValidRefdes,
-  isGroundNet,
-} from "./circuit-traversal.js";
-import {
-  compactArray,
-  getPinNet,
-  isErrorResult,
-  type ParsedNetlist,
-  type ComponentDetails,
-  type CircuitComponent,
-  type AggregatedCircuitResult,
-  type AggregatedComponent,
-  type ErrorResult,
-  type ComponentGroup,
-  type ListComponentsResult,
-  type ListNetsResult,
-  type SearchComponentsResult,
-  type SearchNetsResult,
-  type QueryComponentResult,
-  type CadenceInstall,
-  type ExportNetlistResult,
-  type DiscoveredDesign,
-} from "./types.js";
-
-// Serialize pstswp invocations to prevent concurrent Cadence license conflicts
-const serializePstswp = createMutex();
-
-// =============================================================================
-// Design Loading
-// =============================================================================
-
-/**
- * Normalize unconnected pins to "NC" (No Connect).
- */
-const normalizeUnconnectedPins = (netlist: ParsedNetlist): void => {
-  for (const component of Object.values(netlist.components)) {
-    for (const [pin, net] of Object.entries(component.pins)) {
-      if (typeof net === "string") {
-        if (net === "") {
-          component.pins[pin] = "NC";
-        }
-        continue;
-      }
-
-      if (net?.net === "") {
-        net.net = "NC";
-      }
-    }
-  }
-};
-
-/**
- * Load netlist from a design file path.
- * Delegates to the appropriate handler based on file extension.
- */
-export const loadNetlist = async (designPath: string): Promise<ParsedNetlist | ErrorResult> => {
-  const normalizedPath = resolvePath(designPath);
-  const handler = findHandler(normalizedPath);
-  if (!handler) {
-    const ext = path.extname(normalizedPath);
-    return {
-      error: `Unsupported design file format '${ext}'. Supported: .dsn, .cpm (Cadence), .PrjPcb, .SchDoc (Altium)`,
-    };
-  }
-
-  try {
-    const parsed = await parseDesign(normalizedPath);
-    normalizeUnconnectedPins(parsed);
-    return parsed;
-  } catch (error) {
-    const message = error instanceof Error ? error.message : "Unknown error occurred";
-    return { error: message };
-  }
-};
-
-// =============================================================================
-// Component Grouping
-// =============================================================================
-
-const MPN_MISSING_NOTE =
-  "MPN not found in exported netlist data. Tell user to update symbol properties in library, or to point you to the BOM";
-
-/**
- * Group components by MPN for compact output.
- */
-const groupComponentsByMpn = (
-  entries: Array<[string, ComponentDetails[string]]>,
-  includeDns: boolean
-): ComponentGroup[] => {
-  const groups = new Map<
-    string,
-    {
-      mpn?: string;
-      description?: string;
-      comment?: string;
-      value?: string;
-      dns?: boolean;
-      notes?: string[];
-      refdes: string[];
-    }
-  >();
-
-  for (const [refdes, component] of entries) {
-    const dns = component.dns ?? false;
-    if (!includeDns && dns) {
-      continue;
-    }
-
-    const mpnTrimmed = component.mpn?.trim() || undefined;
-    const descriptionValue = component.description?.trim() || undefined;
-    const commentValue = component.comment?.trim() || undefined;
-    const valueValue = component.value?.trim() || undefined;
-
-    const keyBase = mpnTrimmed ? `mpn:${mpnTrimmed}` : `refdes:${refdes}`;
-    const groupKey = `${keyBase}||dns:${dns ? "1" : "0"}`;
-
-    if (!groups.has(groupKey)) {
-      groups.set(groupKey, {
-        ...(mpnTrimmed && { mpn: mpnTrimmed }),
-        description: descriptionValue,
-        comment: commentValue,
-        value: valueValue,
-        dns: dns || undefined,
-        notes: mpnTrimmed ? undefined : [MPN_MISSING_NOTE],
-        refdes: [],
-      });
-    } else if (valueValue && !groups.get(groupKey)!.value) {
-      groups.get(groupKey)!.value = valueValue;
-    }
-
-    groups.get(groupKey)!.refdes.push(refdes);
-  }
-
-  return Array.from(groups.values())
-    .map((group) => {
-      const entry: ComponentGroup = {
-        count: group.refdes.length,
-        refdes: compactArray(group.refdes.sort(naturalSort)),
-      };
-
-      if (group.mpn !== undefined) {
-        entry.mpn = group.mpn;
-      }
-
-      if (group.description !== undefined) {
-        entry.description = group.description;
-      }
-
-      if (group.comment !== undefined) {
-        entry.comment = group.comment;
-      }
-
-      if (group.value !== undefined) {
-        entry.value = group.value;
-      }
-
-      if (group.dns !== undefined) {
-        entry.dns = group.dns;
-      }
-
-      if (group.notes !== undefined) {
-        entry.notes = group.notes;
-      }
-
-      return entry;
-    })
-    .sort((a, b) => (a.mpn ?? "").localeCompare(b.mpn ?? ""));
-};
-
-/**
- * Aggregate circuit components by MPN for compact output.
- */
-const aggregateCircuitByMpn = (
-  components: CircuitComponent[]
-): AggregatedCircuitResult["components_by_mpn"] => {
-  const groups = new Map<
-    string,
-    {
-      mpn?: string;
-      description?: string;
-      comment?: string;
-      value?: string;
-      dns?: boolean;
-      notes?: string[];
-      orientations: Map<
-        string,
-        {
-          count: number;
-          refdes: string[];
-          connections: Array<{ net: string; pins: string[] }>;
-        }
-      >;
-    }
-  >();
-
-  const unaggregatable: typeof components = [];
-
-  for (const comp of components) {
-    const mpn = comp.mpn?.trim() || undefined;
-    const description = comp.description?.trim() || "";
-    const value = comp.value?.trim() || undefined;
-    const dnsFlag = comp.dns ? true : undefined;
-
-    let aggregationKey: string;
-    if (mpn) {
-      aggregationKey = `mpn:${mpn}`;
-    } else if (description) {
-      aggregationKey = `desc:${description}`;
-    } else {
-      unaggregatable.push(comp);
-      continue;
-    }
-
-    const nets = comp.connections.map((p) => p.net);
-    const netPair = [...nets].sort().join("|");
-    const groupKey = `${aggregationKey}||${netPair}||dns:${dnsFlag ? "1" : "0"}`;
-
-    if (!groups.has(groupKey)) {
-      groups.set(groupKey, {
-        ...(mpn && { mpn }),
-        description: description || undefined,
-        comment: comp.comment,
-        value,
-        dns: dnsFlag,
-        notes: mpn ? undefined : [MPN_MISSING_NOTE],
-        orientations: new Map(),
-      });
-    } else if (value && !groups.get(groupKey)!.value) {
-      groups.get(groupKey)!.value = value;
-    }
-
-    const orientationKey = comp.connections.map((p) => `${p.pins.join(",")}:${p.net}`).join("|");
-    const group = groups.get(groupKey)!;
-
-    if (!group.orientations.has(orientationKey)) {
-      group.orientations.set(orientationKey, {
-        count: 0,
-        refdes: [],
-        connections: comp.connections,
-      });
-    }
-
-    const orientation = group.orientations.get(orientationKey)!;
-    orientation.count++;
-    if (comp.refdes) {
-      orientation.refdes.push(comp.refdes);
-    }
-  }
-
-  const compactConnections = (connections: Array<{ net: string; pins: string[] }>) =>
-    connections.map((c) => ({ net: c.net, pins: compactArray(c.pins) }));
-
-  const result: AggregatedComponent[] = [];
-
-  for (const group of groups.values()) {
-    const orientationsList = Array.from(group.orientations.values()).sort(
-      (a, b) => b.count - a.count
-    );
-
-    const totalCount = orientationsList.reduce((sum, o) => sum + o.count, 0);
-
-    const aggregated: AggregatedComponent = {
-      total_count: totalCount,
-    };
-
-    if (group.mpn !== undefined) {
-      aggregated.mpn = group.mpn;
-    }
-
-    if (group.description !== undefined) {
-      aggregated.description = group.description;
-    }
-    if (group.comment !== undefined) {
-      aggregated.comment = group.comment;
-    }
-    if (group.value !== undefined) {
-      aggregated.value = group.value;
-    }
-    if (group.dns !== undefined) {
-      aggregated.dns = group.dns;
-    }
-    if (group.notes !== undefined) {
-      aggregated.notes = group.notes;
-    }
-
-    if (orientationsList.length === 1) {
-      aggregated.refdes = compactArray(orientationsList[0].refdes.sort(naturalSort));
-      aggregated.connections = compactConnections(orientationsList[0].connections);
-    } else {
-      aggregated.orientations = orientationsList.map((o) => ({
-        count: o.count,
-        refdes: compactArray(o.refdes.sort(naturalSort)),
-        connections: compactConnections(o.connections),
-      }));
-    }
-
-    result.push(aggregated);
-  }
-
-  for (const comp of unaggregatable) {
-    const unagg: AggregatedComponent = {
-      refdes: comp.refdes,
-      notes: [MPN_MISSING_NOTE],
-      total_count: 1,
-      connections: compactConnections(comp.connections),
-    };
-
-    if (comp.description !== undefined) {
-      unagg.description = comp.description;
-    }
-    if (comp.comment !== undefined) {
-      unagg.comment = comp.comment;
-    }
-    if (comp.value !== undefined) {
-      unagg.value = comp.value;
-    }
-    if (comp.dns) {
-      unagg.dns = true;
-    }
-
-    result.push(unagg);
-  }
-
-  return result.sort((a, b) => b.total_count - a.total_count);
-};
-
-// =============================================================================
-// Regex Helpers
-// =============================================================================
-
-/**
- * Parse PCRE-style inline flags from a regex pattern and convert to JS RegExp flags.
- * Supports (?i), (?m), (?s), (?u) and combinations like (?im) at the start of the pattern.
- */
-const parseRegexPattern = (
-  pattern: string,
-  defaultFlags = ""
-): { regex: RegExp } | { error: string } => {
-  let flags = defaultFlags;
-  let cleanPattern = pattern;
-
-  const inlineFlagMatch = cleanPattern.match(/^\(\?([imsu]+)\)/);
-  if (inlineFlagMatch) {
-    cleanPattern = cleanPattern.slice(inlineFlagMatch[0].length);
-    const allFlags = new Set([...flags, ...inlineFlagMatch[1]]);
-    flags = [...allFlags].filter((f) => "gimsuvy".includes(f)).join("");
-  }
-
-  try {
-    return { regex: new RegExp(cleanPattern, flags) };
-  } catch {
-    return { error: `Invalid regex pattern '${pattern}'` };
-  }
-};
-
-/**
- * Return an error when a search pattern matches every item in the dataset.
- * This prevents wildcard patterns (e.g. `.*`) from dumping the full list.
- */
-const tooManyMatchesError = (
-  pattern: string,
-  matchCount: number,
-  toolSuggestion: string
-): ErrorResult => ({
-  error: `Pattern '${pattern}' matched all ${matchCount} items. Use ${toolSuggestion} to retrieve the full list, or use a more specific pattern.`,
-});
-
-// =============================================================================
-// Public API
-// =============================================================================
-
-/**
- * Options for listDesigns.
- */
-export interface ListDesignsOptions {
-  searchPath?: string;
-  pattern?: string;
-  maxDepth?: number;
-  maxResults?: number;
-}
-
-/**
- * Build the path fields for a discovered design.
- * For Cadence CIS/HDL with .dat files: path=pstxnet.dat (preferred), source=.DSN
- * For Cadence CIS/HDL without .dat files: path=.DSN
- * For all others: path=sourcePath
- */
-const getDesignPaths = (design: DiscoveredDesign): { path: string; source?: string } => {
-  if (design.format === "cadence-cis" || design.format === "cadence-hdl") {
-    if (design.datFiles.pstxnet) {
-      return { path: design.datFiles.pstxnet, source: design.sourcePath };
-    }
-  }
-  return { path: design.sourcePath };
-};
-
-/**
- * List all designs in a directory.
- */
-export const listDesigns = async (
-  options: ListDesignsOptions = {}
-): Promise<
-  Array<{ name: string; path: string; source?: string; error?: string }> | ErrorResult
-> => {
-  const { searchPath, pattern = ".*", maxDepth, maxResults = 50 } = options;
-  const resolvedPath = resolvePath(searchPath ?? ".");
-
-  const parsed = parseRegexPattern(pattern);
-  if ("error" in parsed) return parsed;
-  const regex = parsed.regex;
-
-  let designs;
-  try {
-    designs = await discoverDesigns(resolvedPath, { maxDepth });
-  } catch (error) {
-    const message = error instanceof Error ? error.message : "Unknown error occurred";
-    return { error: `Failed to search '${resolvedPath}': ${message}` };
-  }
-
-  const filtered = designs.filter((design) => regex.test(design.name));
-  const limited = filtered.slice(0, maxResults);
-  return limited.map((design) => ({
-    name: design.name,
-    ...getDesignPaths(design),
-    error: design.error,
-  }));
-};
-
-/**
- * List components of a specific type in a design.
- *
- * @param design - Path to design file
- * @param type - Component type prefix (e.g., "U", "R", "C")
- * @param includeDns - Include DNS (Do Not Stuff) components
- */
-export const listComponents = async (
-  design: string,
-  type: string,
-  includeDns = false
-): Promise<ListComponentsResult | ErrorResult> => {
-  const netlist = await loadNetlist(design);
-  if (isErrorResult(netlist)) {
-    return netlist;
-  }
-
-  const prefix = type.trim().toUpperCase();
-  if (!prefix) {
-    return { error: "Missing required parameter: type" };
-  }
-
-  const entries = Object.entries(netlist.components).filter(([refdes]) =>
-    matchesRefdesType(refdes, prefix)
-  );
-
-  if (entries.length === 0) {
-    const availablePrefixes = Array.from(
-      new Set(Object.keys(netlist.components).filter(isValidRefdes).map(getRefdesPrefix))
-    ).sort((a, b) => a.localeCompare(b));
-
-    const designName = path.basename(design, path.extname(design));
-    return {
-      error: `No components with prefix '${prefix}' found in design '${designName}'. Available prefixes: [${availablePrefixes.join(", ")}]`,
-    };
-  }
-
-  return {
-    components: groupComponentsByMpn(entries, includeDns),
-  };
-};
-
-/**
- * List all nets within a design.
- *
- * @param design - Path to design file
- */
-export const listNets = async (design: string): Promise<ListNetsResult | ErrorResult> => {
-  const netlist = await loadNetlist(design);
-  if (isErrorResult(netlist)) {
-    return netlist;
-  }
-
-  const nets = Object.keys(netlist.nets).sort((a, b) => a.localeCompare(b));
-  return { nets };
-};
-
-/**
- * Search nets by regex pattern.
- *
- * @param pattern - Regex pattern
- * @param design - Path to design file
- */
-export const searchNets = async (
-  pattern: string,
-  design: string
-): Promise<SearchNetsResult | ErrorResult> => {
-  const parsed = parseRegexPattern(pattern, "i");
-  if ("error" in parsed) return parsed;
-  const regex = parsed.regex;
-
-  const netlist = await loadNetlist(design);
-  if (isErrorResult(netlist)) {
-    return netlist;
-  }
-
-  const designName = path.basename(design, path.extname(design));
-  const allNets = Object.keys(netlist.nets);
-  const nets = allNets.filter((net) => regex.test(net));
-
-  if (nets.length > 0 && nets.length === allNets.length) {
-    return tooManyMatchesError(pattern, nets.length, "list_nets");
-  }
-
-  const sorted = nets.sort((a, b) => a.localeCompare(b));
-
-  if (sorted.length === 0) {
-    return {
-      results: { [designName]: [] },
-      notes: [`No nets matched pattern '${pattern}'`],
-    };
-  }
-
-  return { results: { [designName]: sorted } };
-};
-
-/**
- * Search components by refdes pattern.
- *
- * @param pattern - Regex pattern
- * @param design - Path to design file
- * @param includeDns - Include DNS components
- */
-export const searchComponentsByRefdes = async (
-  pattern: string,
-  design: string,
-  includeDns = false
-): Promise<SearchComponentsResult | ErrorResult> => {
-  const parsed = parseRegexPattern(pattern, "i");
-  if ("error" in parsed) return parsed;
-  const regex = parsed.regex;
-
-  const netlist = await loadNetlist(design);
-  if (isErrorResult(netlist)) {
-    return netlist;
-  }
-
-  const designName = path.basename(design, path.extname(design));
-  const allEntries = Object.entries(netlist.components);
-  const entries = allEntries.filter(([refdes]) => regex.test(refdes));
-
-  if (entries.length > 0 && entries.length === allEntries.length) {
-    return tooManyMatchesError(pattern, entries.length, "list_components");
-  }
-
-  const grouped = groupComponentsByMpn(entries, includeDns);
-
-  if (grouped.length === 0) {
-    return {
-      results: { [designName]: [] },
-      notes: [`No components matched refdes pattern '${pattern}'`],
-    };
-  }
-
-  return { results: { [designName]: grouped } };
-};
-
-/**
- * Search components by MPN pattern.
- *
- * @param pattern - Regex pattern
- * @param design - Path to design file
- * @param includeDns - Include DNS components
- */
-export const searchComponentsByMpn = async (
-  pattern: string,
-  design: string,
-  includeDns = false
-): Promise<SearchComponentsResult | ErrorResult> => {
-  const parsed = parseRegexPattern(pattern, "i");
-  if ("error" in parsed) return parsed;
-  const regex = parsed.regex;
-
-  const netlist = await loadNetlist(design);
-  if (isErrorResult(netlist)) {
-    return netlist;
-  }
-
-  const designName = path.basename(design, path.extname(design));
-  const allComponents = Object.entries(netlist.components);
-  const componentsWithMpn = allComponents.filter(([, c]) => c.mpn?.trim());
-  const entries = componentsWithMpn.filter(([, component]) => regex.test(component.mpn!));
-
-  // Case 1: No MPN data exists at all
-  if (componentsWithMpn.length === 0) {
-    return {
-      results: { [designName]: [] },
-      notes: ["This netlist has no MPN data. Ask user for BOM or schematic PDF"],
-    };
-  }
-
-  if (entries.length > 0 && entries.length === componentsWithMpn.length) {
-    return tooManyMatchesError(pattern, entries.length, "list_components");
-  }
-
-  const grouped = groupComponentsByMpn(entries, includeDns);
-
-  // Case 2: MPN data exists but pattern didn't match
-  if (grouped.length === 0) {
-    return {
-      results: { [designName]: [] },
-      notes: [
-        `No components matched pattern '${pattern}'. Try a broader pattern or use search_components_by_refdes instead`,
-      ],
-    };
-  }
-
-  return { results: { [designName]: grouped } };
-};
-
-/**
- * Search components by description pattern.
- *
- * @param pattern - Regex pattern
- * @param design - Path to design file
- * @param includeDns - Include DNS components
- */
-export const searchComponentsByDescription = async (
-  pattern: string,
-  design: string,
-  includeDns = false
-): Promise<SearchComponentsResult | ErrorResult> => {
-  const parsed = parseRegexPattern(pattern, "i");
-  if ("error" in parsed) return parsed;
-  const regex = parsed.regex;
-
-  const netlist = await loadNetlist(design);
-  if (isErrorResult(netlist)) {
-    return netlist;
-  }
-
-  const designName = path.basename(design, path.extname(design));
-  const allComponents = Object.entries(netlist.components);
-  const componentsWithDescription = allComponents.filter(([, c]) => c.description?.trim());
-  const entries = componentsWithDescription.filter(([, component]) =>
-    regex.test(component.description!)
-  );
-
-  // Case 1: No description data exists at all
-  if (componentsWithDescription.length === 0) {
-    return {
-      results: { [designName]: [] },
-      notes: ["This netlist has no description data. Ask user for BOM or schematic PDF"],
-    };
-  }
-
-  if (entries.length > 0 && entries.length === componentsWithDescription.length) {
-    return tooManyMatchesError(pattern, entries.length, "list_components");
-  }
-
-  const grouped = groupComponentsByMpn(entries, includeDns);
-
-  // Case 2: Description data exists but pattern didn't match
-  if (grouped.length === 0) {
-    return {
-      results: { [designName]: [] },
-      notes: [
-        `No components matched pattern '${pattern}'. Try a broader pattern or use search_components_by_refdes instead`,
-      ],
-    };
-  }
-
-  return { results: { [designName]: grouped } };
-};
-
-/**
- * Query component details by reference designator.
- *
- * @param design - Path to design file
- * @param refdes - Component reference designator
- */
-export const queryComponent = async (
-  design: string,
-  refdes: string
-): Promise<QueryComponentResult | ErrorResult> => {
-  const netlist = await loadNetlist(design);
-  if (isErrorResult(netlist)) {
-    return netlist;
-  }
-
-  const targetRefdes = refdes.trim();
-  const componentEntry = Object.entries(netlist.components).find(
-    ([key]) => key.toLowerCase() === targetRefdes.toLowerCase()
-  );
-
-  if (!componentEntry) {
-    const designName = path.basename(design, path.extname(design));
-    return {
-      error: `Component '${refdes}' not found in design '${designName}'. Use list_components() or search_components_by_refdes() to find available components.`,
-    };
-  }
-
-  const [resolvedRefdes, component] = componentEntry;
-  const mpn = component.mpn?.trim() || undefined;
-  const dns = component.dns ?? false;
-
-  const result: QueryComponentResult = {
-    refdes: resolvedRefdes,
-    pins: component.pins,
-  };
-
-  if (mpn !== undefined) {
-    result.mpn = mpn;
-  }
-
-  if (component.description !== undefined) {
-    result.description = component.description;
-  }
-  if (component.comment !== undefined) {
-    result.comment = component.comment;
-  }
-  if (component.value !== undefined) {
-    result.value = component.value;
-  }
-  if (dns) {
-    result.dns = true;
-  }
-  if (!mpn) {
-    result.notes = [MPN_MISSING_NOTE];
-  }
-
-  return result;
-};
-
-/**
- * Query circuit starting from a net name.
- *
- * @param design - Path to design file
- * @param netName - Net name
- * @param skipTypes - Component types to skip
- * @param includeDns - Include DNS components
- */
-export const queryXnetByNetName = async (
-  design: string,
-  netName: string,
-  skipTypes: string[] = [],
-  includeDns = false
-): Promise<AggregatedCircuitResult | ErrorResult> => {
-  const netlist = await loadNetlist(design);
-  if (isErrorResult(netlist)) {
-    return netlist;
-  }
-
-  const { nets, components } = netlist;
-
-  if (!nets[netName]) {
-    const designName = path.basename(design, path.extname(design));
-    return {
-      error: `Net '${netName}' not found in design '${designName}'. Use search_nets() to find available nets.`,
-    };
-  }
-
-  if (isGroundNet(netName)) {
-    return {
-      error: `${netName} is a ground net and cannot be queried.`,
-    };
-  }
-
-  const traversal = traverseCircuitFromNet(netName, nets, components, {
-    skipTypes,
-    includeDns,
-  });
-
-  const circuitHash = computeCircuitHash(traversal.components);
-  const aggregated = aggregateCircuitByMpn(traversal.components);
-
-  const response: AggregatedCircuitResult = {
-    starting_point: netName,
-    total_components: traversal.components.length,
-    unique_configurations: aggregated.length,
-    components_by_mpn: aggregated,
-    visited_nets: traversal.visited_nets,
-    circuit_hash: circuitHash,
-  };
-
-  if (Object.keys(traversal.skipped).length > 0) {
-    response.skipped = traversal.skipped;
-  }
-
-  return response;
-};
-
-/**
- * Query circuit starting from a component pin.
- *
- * @param design - Path to design file
- * @param pinSpec - Pin specification in "REFDES.PIN" format
- * @param skipTypes - Component types to skip
- * @param includeDns - Include DNS components
- */
-export const queryXnetByPinName = async (
-  design: string,
-  pinSpec: string,
-  skipTypes: string[] = [],
-  includeDns = false
-): Promise<AggregatedCircuitResult | ErrorResult> => {
-  const netlist = await loadNetlist(design);
-  if (isErrorResult(netlist)) {
-    return netlist;
-  }
-
-  const parts = pinSpec.split(".");
-  if (parts.length !== 2) {
-    return {
-      error: `Invalid pin name '${pinSpec}'. Expected 'REFDES.PIN'.`,
-    };
-  }
-
-  const [refdesInput, pinInput] = parts;
-  const refdesEntry = Object.entries(netlist.components).find(
-    ([refdes]) => refdes.toLowerCase() === refdesInput.trim().toLowerCase()
-  );
-
-  if (!refdesEntry) {
-    const designName = path.basename(design, path.extname(design));
-    return {
-      error: `Component '${refdesInput}' not found in design '${designName}'. Use list_components() or search_components_by_refdes() to find available components.`,
-    };
-  }
-
-  const [resolvedRefdes, component] = refdesEntry;
-  const pinKey = Object.keys(component.pins).find(
-    (pin) => pin.toLowerCase() === pinInput.trim().toLowerCase()
-  );
-
-  if (!pinKey) {
-    const pins = Object.keys(component.pins).sort(naturalSort);
-    return {
-      error: `Pin '${pinSpec}' not found. Component ${resolvedRefdes} has pins: [${pins.join(", ")}]`,
-    };
-  }
-
-  const connectedNet = getPinNet(component.pins[pinKey]);
-
-  if (isGroundNet(connectedNet)) {
-    return {
-      error: `Pin ${resolvedRefdes}.${pinKey} is connected to ${connectedNet} (ground) and cannot be queried.`,
-    };
-  }
-
-  if (connectedNet === "NC") {
-    return {
-      starting_point: `${resolvedRefdes}.${pinKey}`,
-      net: "NC",
-      total_components: 0,
-      unique_configurations: 0,
-      components_by_mpn: [],
-      visited_nets: ["NC"],
-      circuit_hash: `nc-${resolvedRefdes}.${pinKey}`,
-    };
-  }
-
-  const { nets, components } = netlist;
-  const traversal = traverseCircuitFromNet(connectedNet, nets, components, {
-    skipTypes,
-    includeDns,
-  });
-
-  const circuitHash = computeCircuitHash(traversal.components);
-  const aggregated = aggregateCircuitByMpn(traversal.components);
-
-  const response: AggregatedCircuitResult = {
-    starting_point: `${resolvedRefdes}.${pinKey}`,
-    net: connectedNet,
-    total_components: traversal.components.length,
-    unique_configurations: aggregated.length,
-    components_by_mpn: aggregated,
-    visited_nets: traversal.visited_nets,
-    circuit_hash: circuitHash,
-  };
-
-  if (Object.keys(traversal.skipped).length > 0) {
-    response.skipped = traversal.skipped;
-  }
-
-  return response;
-};
-
-// =============================================================================
-// Cadence Netlist Export (Windows Only)
-// =============================================================================
-
-const execAsync = promisify(exec);
-
-/**
- * Convert Windows path to bash-compatible path for GitBash/WSL compatibility.
- * Example: C:\foo\bar -> /c/foo/bar
- */
-const toBashPath = (winPath: string): string =>
-  winPath
-    .replace(/\\/g, "/")
-    .replace(/^([A-Za-z]):/, (_, drive: string) => `/${drive.toLowerCase()}`);
-
-/**
- * Detect installed Cadence SPB versions from the standard installation directory.
- *
- * @param cadenceBase - Base Cadence installation directory (default: C:/Cadence)
- * @returns Array of detected Cadence installations, sorted by version descending
- */
-export const detectCadenceVersions = async (
-  cadenceBase = "C:/Cadence"
-): Promise<CadenceInstall[]> => {
-  const installs: CadenceInstall[] = [];
-
-  try {
-    const entries = await fs.promises.readdir(cadenceBase);
-
-    for (const entry of entries) {
-      const match = entry.match(/^SPB_(\d+\.\d+)$/);
-      if (!match) continue;
-
-      const version = match[1];
-      const root = path.join(cadenceBase, entry);
-      const pstswp = path.join(root, "tools", "bin", "pstswp.exe");
-      const config = path.join(root, "tools", "capture", "allegro.cfg");
-
-      // Verify the executables exist
-      if (fs.existsSync(pstswp) && fs.existsSync(config)) {
-        installs.push({ version, root, pstswp, config });
-      }
-    }
-
-    // Sort by version descending (latest first)
-    installs.sort((a, b) => parseFloat(b.version) - parseFloat(a.version));
-  } catch {
-    // Cadence directory doesn't exist or isn't accessible
-  }
-
-  return installs;
-};
-
-/**
- * Get the latest installed Cadence version.
- *
- * @returns The latest Cadence installation, or null if none found
- */
-export const getLatestCadence = async (): Promise<CadenceInstall | null> => {
-  const versions = await detectCadenceVersions();
-  return versions[0] ?? null;
-};
-
-/**
- * Resolve the Allegro output directory for netlist export.
- * Uses an existing Allegro/ or allegro/ directory if present, otherwise creates allegro/.
- */
-export const resolveAllegroDir = async (
-  dsnDir: string
-): Promise<{ outputDir: string; dirName: string }> => {
-  let dirName = "allegro";
-  try {
-    const entries = await fs.promises.readdir(dsnDir);
-    for (const candidate of ["Allegro", "allegro"]) {
-      if (entries.includes(candidate)) {
-        dirName = candidate;
-        break;
-      }
-    }
-  } catch {
-    // parent doesn't exist or can't be read
-  }
-  const outputDir = path.join(dsnDir, dirName);
-  await fs.promises.mkdir(outputDir, { recursive: true });
-  return { outputDir, dirName };
-};
-
-/**
- * Temporarily relocate a .DSNlck lock file so pstswp can proceed.
- * Returns the temporary path if relocated, or undefined if no lock file exists.
- */
-export const relocateLockFile = async (dsnPath: string): Promise<string | undefined> => {
-  const lockPath = dsnPath.replace(/\.DSN$/i, ".DSNlck");
-  try {
-    await fs.promises.access(lockPath);
-  } catch {
-    return undefined;
-  }
-  const tempPath = path.join(tmpdir(), `${path.basename(lockPath)}.${Date.now()}`);
-  await fs.promises.rename(lockPath, tempPath);
-  return tempPath;
-};
-
-/**
- * Restore a previously relocated .DSNlck lock file.
- * Logs a warning if restoration fails (e.g. temp file was cleaned up).
- */
-export const restoreLockFile = async (dsnPath: string, tempPath: string): Promise<void> => {
-  const lockPath = dsnPath.replace(/\.DSN$/i, ".DSNlck");
-  try {
-    await fs.promises.rename(tempPath, lockPath);
-  } catch {
-    console.warn(`Failed to restore lock file. Temporary location: ${tempPath}`);
-  }
-};
-
-/**
- * Export Cadence schematic netlist to Allegro PCB format.
- * Uses the pstswp utility from Cadence SPB installation.
- *
- * @param dsnPath - Path to .DSN schematic file
- * @returns Export result with output directory and generated files, or error
- */
-export const exportCadenceNetlist = async (
-  dsnPath: string
-): Promise<ExportNetlistResult | ErrorResult> => {
-  // Platform check
-  if (process.platform !== "win32") {
-    return {
-      error:
-        "Cadence export tools are only available on Windows. The pstswp utility requires a Windows environment with Cadence SPB installed. Manual export: Open Cadence, then: Tools → Create Netlist → PCB Editor format.",
-    };
-  }
-
-  // Find Cadence installation
-  const cadence = await getLatestCadence();
-  if (!cadence) {
-    return {
-      error:
-        "No Cadence SPB installation found in C:/Cadence. Ensure Cadence Design Entry CIS or HDL is installed. Manual export: Open Cadence, then: Tools → Create Netlist → PCB Editor format.",
-    };
-  }
-
-  const resolvedDsnPath = resolvePath(dsnPath);
-  const dsnDir = path.dirname(resolvedDsnPath);
-  const dsnFile = path.basename(dsnPath);
-  const { outputDir, dirName: outputDirName } = await resolveAllegroDir(dsnDir);
-
-  return serializePstswp(async () => {
-    // Temporarily relocate .DSNlck lock file if present (stale locks block pstswp)
-    const lockTempPath = await relocateLockFile(resolvedDsnPath);
-
-    // Convert to bash paths for command execution (GitBash compatibility)
-    const bashDsnDir = toBashPath(dsnDir);
-    const pstswp = toBashPath(cadence.pstswp);
-    const config = toBashPath(cadence.config);
-
-    const command = `cd "${bashDsnDir}" && "${pstswp}" -pst -d "${dsnFile}" -n "${outputDirName}" -c "${config}" -v 3 -l 255 -j "PCB Footprint"`;
-
-    try {
-      const { stdout, stderr } = await execAsync(command, {
-        shell: "bash",
-        timeout: 120000,
-      });
-
-      // List generated files
-      let generatedFiles: string[] | undefined;
-      try {
-        const files = await fs.promises.readdir(outputDir);
-        generatedFiles = files.sort();
-      } catch {
-        // Output directory may not exist if export failed silently
-      }
-
-      return {
-        success: true,
-        outputDir,
-        log: (stdout + stderr).trim() || undefined,
-        cadenceVersion: cadence.version,
-        generatedFiles,
-      };
-    } catch (err: unknown) {
-      const execError = err as {
-        message?: string;
-        stdout?: string;
-        stderr?: string;
-      };
-      const lockNote = lockTempPath
-        ? ` A .DSNlck lock file was found and temporarily relocated — this is often the cause of pstswp failures.`
-        : "";
-      return {
-        error: `Cadence pstswp failed: ${execError.message ?? "Unknown error"}${lockNote}`,
-      };
-    } finally {
-      if (lockTempPath) {
-        await restoreLockFile(resolvedDsnPath, lockTempPath);
-      }
-    }
-  });
-};
-
-// Re-export types that appear in public API signatures
-export type {
-  ParsedNetlist,
-  ErrorResult,
-  ListComponentsResult,
-  ListNetsResult,
-  SearchComponentsResult,
-  SearchNetsResult,
-  QueryComponentResult,
-  AggregatedCircuitResult,
-  CadenceInstall,
-  ExportNetlistResult,
-  ComponentGroup,
-  AggregatedComponent,
-  ComponentDetails,
-  CircuitComponent,
-  PinEntry,
-} from "./types.js";
-
-export { isErrorResult } from "./types.js";
-
-// =============================================================================
-// Test Exports
-// =============================================================================
-
-/**
- * Internal exports for testing purposes only.
- * @internal
- */
-export { MPN_MISSING_NOTE, groupComponentsByMpn, aggregateCircuitByMpn, parseRegexPattern };
