+++ src/service/tools/search-nets.test.ts
@@ -1,0 +1,118 @@
+import { describe, it, expect, vi, beforeEach, afterEach } from "vitest";
+import { searchNets } from "./search-nets.js";
+import type { ParsedNetlist, ErrorResult } from "../../types.js";
+import * as parsersModule from "../../parsers/index.js";
+
+const isErrorResult = (result: unknown): result is ErrorResult =>
+  typeof result === "object" && result !== null && "error" in result;
+
+describe("searchNets - case insensitive by default", () => {
+  beforeEach(() => {
+    const mockNetlist: ParsedNetlist = {
+      nets: {
+        VDD_1V8: { U1: "1" },
+        VDD_3V3: { U2: "1" },
+        GND: { U1: "2", U2: "2" },
+      },
+      components: {
+        U1: { pins: { "1": "VDD_1V8", "2": "GND" } },
+        U2: { pins: { "1": "VDD_3V3", "2": "GND" } },
+      },
+    };
+    vi.spyOn(parsersModule, "parseDesign").mockResolvedValue(mockNetlist);
+  });
+
+  afterEach(() => {
+    vi.restoreAllMocks();
+  });
+
+  it("lowercase pattern matches uppercase net names", async () => {
+    const result = await searchNets("vdd", "/mock/design.dsn");
+    expect(isErrorResult(result)).toBe(false);
+    if (!isErrorResult(result)) {
+      const nets = Object.values(result.results).flat();
+      expect(nets).toContain("VDD_1V8");
+      expect(nets).toContain("VDD_3V3");
+    }
+  });
+
+  it("uppercase pattern still works", async () => {
+    const result = await searchNets("VDD", "/mock/design.dsn");
+    expect(isErrorResult(result)).toBe(false);
+    if (!isErrorResult(result)) {
+      const nets = Object.values(result.results).flat();
+      expect(nets).toContain("VDD_1V8");
+      expect(nets).toContain("VDD_3V3");
+    }
+  });
+
+  it("explicit (?i) flag also works", async () => {
+    const result = await searchNets("(?i)gnd", "/mock/design.dsn");
+    expect(isErrorResult(result)).toBe(false);
+    if (!isErrorResult(result)) {
+      const nets = Object.values(result.results).flat();
+      expect(nets).toContain("GND");
+    }
+  });
+});
+
+describe("searchNets - broad pattern rejection", () => {
+  const mockNetlist: ParsedNetlist = {
+    nets: {
+      VDD_1V8: { U1: "1" },
+      GND: { U1: "2", R1: "2" },
+      SIG_A: { R1: "1" },
+    },
+    components: {
+      U1: {
+        mpn: "TPS62088",
+        description: "Buck Converter",
+        pins: { "1": "VDD_1V8", "2": "GND" },
+      },
+      R1: {
+        mpn: "RC0402FR-0710KL",
+        description: "10K Resistor",
+        pins: { "1": "SIG_A", "2": "GND" },
+      },
+    },
+  };
+
+  beforeEach(() => {
+    vi.spyOn(parsersModule, "findHandler").mockReturnValue({
+      name: "mock",
+      extensions: [".dsn"],
+      canHandle: () => true,
+      discoverDesigns: vi.fn(),
+      parse: vi.fn(),
+    });
+    vi.spyOn(parsersModule, "parseDesign").mockResolvedValue(mockNetlist);
+  });
+
+  afterEach(() => {
+    vi.restoreAllMocks();
+  });
+
+  it("rejects .* pattern that matches all nets", async () => {
+    const result = await searchNets(".*", "/mock/design.dsn");
+    expect(isErrorResult(result)).toBe(true);
+    expect((result as ErrorResult).error).toContain("list_nets");
+    expect((result as ErrorResult).error).toContain("all 3 items");
+  });
+
+  it("rejects .+ pattern that matches all nets", async () => {
+    const result = await searchNets(".+", "/mock/design.dsn");
+    expect(isErrorResult(result)).toBe(true);
+    expect((result as ErrorResult).error).toContain("list_nets");
+  });
+
+  it("rejects ^ pattern that matches all nets", async () => {
+    const result = await searchNets("^", "/mock/design.dsn");
+    expect(isErrorResult(result)).toBe(true);
+    expect((result as ErrorResult).error).toContain("list_nets");
+  });
+
+  it("allows specific pattern that matches subset", async () => {
+    const result = await searchNets("VDD", "/mock/design.dsn");
+    expect(isErrorResult(result)).toBe(false);
+  });
+});
